Extract auto_start tests to separate file
Move the auto_start_decision module from unit.rs to a separate file
at unit/auto_start.rs to comply with the 400-line file limit. The
extraction reduces unit.rs from 401 to 337 lines whilst maintaining
test organisation.
